+++ studio/CHANGELOG.md
@@ -4,6 +4,30 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.73.2](https://github.com/wundergraph/cosmo/compare/studio@0.73.1...studio@0.73.2) (2024-03-18)
+
+### Bug Fixes
+
+* graph overview ui visuals ([#633](https://github.com/wundergraph/cosmo/issues/633)) ([96867e6](https://github.com/wundergraph/cosmo/commit/96867e6fb908c17e02997babdf13e0a5b15515c7)) (@StarpTech)
+
+## [0.73.1](https://github.com/wundergraph/cosmo/compare/studio@0.73.0...studio@0.73.1) (2024-03-16)
+
+### Bug Fixes
+
+* load monaco from public folder ([#641](https://github.com/wundergraph/cosmo/issues/641)) ([a691c95](https://github.com/wundergraph/cosmo/commit/a691c95a6cf2641cc683eba897fa5344ba41bd0c)) (@StarpTech)
+
+# [0.73.0](https://github.com/wundergraph/cosmo/compare/studio@0.72.0...studio@0.73.0) (2024-03-16)
+
+### Bug Fixes
+
+* cookie race ([#636](https://github.com/wundergraph/cosmo/issues/636)) ([6d2baed](https://github.com/wundergraph/cosmo/commit/6d2baed692dd876abae4a984ee39ffa1d39ed4ed)) (@thisisnithin)
+* csp header ([a1ec190](https://github.com/wundergraph/cosmo/commit/a1ec1902eb4aa7af461ee76b63fd8bda3691c0cf)) (@StarpTech)
+* improve ui and logs ([#631](https://github.com/wundergraph/cosmo/issues/631)) ([83695b9](https://github.com/wundergraph/cosmo/commit/83695b97d3aca66d70677bc207f874c8aa17bf65)) (@JivusAyrus)
+
+### Features
+
+* configure CSP for studio ([#639](https://github.com/wundergraph/cosmo/issues/639)) ([183ef61](https://github.com/wundergraph/cosmo/commit/183ef6166bf106ce87d6d7472c9297e7cd63ac31)) (@StarpTech)
+
 # [0.72.0](https://github.com/wundergraph/cosmo/compare/studio@0.71.1...studio@0.72.0) (2024-03-14)
 
 ### Features